% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admit_source_invalid.R
\name{admit_source_invalid}
\alias{admit_source_invalid}
\title{Getting Invalid Examples and Summaries for Admit_Source}
\usage{
admit_source_invalid(data)
}
\arguments{
\item{data}{The raw data from BioSense on which you will do the invalid admit source checks.}
}
\value{
A list of two data frames: examples and summary for invalid Admit_Source.
}
\description{
This function will generate two data frames: first, a frame to be used later to extract invalid examples from;
second, a frame that contains facility-level summaries for counts and percentages of invalid Admit_Source.
}
\details{
The valid admit source value data were taken from the `PHVS_AdmitSource_HL7_2x_V1.xls` file from 
Public Health Information Network Vocabulary Access and Distribution System value sets
(https://phinvads.cdc.gov/vads/ViewView.action?name=Syndromic%20Surveillance). The package will
be updated as the CDC provides new or different codes that are considered valid or invalid.

You can view the concept codes that are considered valid by calling `data("admit_source")`.
}
