% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_pressure_invalid.R
\name{blood_pressure_invalid}
\alias{blood_pressure_invalid}
\title{Getting Invalid Examples and Summaries for for Systolic_Diastolic_Blood_Pressure and Units}
\usage{
blood_pressure_invalid(data)
}
\arguments{
\item{data}{The raw data from BioSense on which you will do the blood pressure checks.}
}
\value{
A list of two data frames: examples and summary for blood pressure checks.
}
\description{
This function will generate two data frames: first, a frame to be used later to extract invalid examples from;
second, a frame that contains facility-level summaries for counts and percentages of Systolic_Diastolic_Blood_Pressure
and Systolic_Diastolic_Blood_Pressure_Units. This function will also mark invalid if Units are given, but Blood_Pressure
is not; it will also return invalid if Blood_Pressure is given, but Units are not. Therefore, it does three total checks:
Are the units invalid? Are units missing, given a blood pressure? Is blood pressure missing, given units?
}
\details{
The valid values were taken from the `PHVS_BloodPressureUnit_UCUM_V1.xls` file from 
Public Health Information Network Vocabulary Access and Distribution System value sets
(https://phinvads.cdc.gov/vads/ViewView.action?name=Syndromic%20Surveillance). The package will
be updated as the CDC provides new or different codes that are considered valid or invalid.

You can view the concept names that are considered valid by calling `data("blood_pressure_units")`.
}
