% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_invalid.R
\name{country_invalid}
\alias{country_invalid}
\title{Getting Invalid Examples and Summaries for Patient_Country}
\usage{
country_invalid(data)
}
\arguments{
\item{data}{The raw data From BioSense on which you will do the invalid Patient_Country check.}
}
\value{
A list of two data frames: examples and summary for Patient_Country.
}
\description{
This function will generate two data frames: first, a frame to be used later to extract invalid examples from;
second, a frame that contains facility-level summaries for counts and percentages of cases with invalid Patient_Country.
}
\details{
The valid values were taken from the `PHVS_Country_ISO_3166-1_V1.xls` file from 
Public Health Information Network Vocabulary Access and Distribution System value sets
(https://phinvads.cdc.gov/vads/ViewView.action?name=Syndromic%20Surveillance). The package will
be updated as the CDC provides new or different codes that are considered valid or invalid.

You can view the concept codes that are considered valid by calling `data("country")`.
}
