% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/death_invalid.R
\name{death_invalid}
\alias{death_invalid}
\title{Checking If Death Date Time, Death Indicator, and Death Discharge Disposition Are Reported Together}
\usage{
death_invalid(data)
}
\arguments{
\item{data}{The raw data from BioSense on which you will do the invalid death checks.}
}
\value{
A list of two data frames: examples and summary for these invalid death checks.
}
\description{
This function will generate two data frames: first, a frame to be used later to extract invalid examples from;
second, a frame that contains facility-level summaries for counts and percentages of death invalids.
If Discharge_Disposition says the patient died, they should report a Death_Indicator AND Death_Date_Time. 
If they do not, this function counts that as invalid. If Death_Indicator is returned, then Death_Date_Time should be reported. 
If they do not, this function counts that as invalid. Death_Indicator should be reported if a Death_Date_Time is reported; 
if it is not, this function counts that as invalid.
}
