% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis_type_invalid.R
\name{diagnosis_type_invalid}
\alias{diagnosis_type_invalid}
\title{Getting Invalid Examples and Summaries for Diagnosis_Type}
\usage{
diagnosis_type_invalid(data)
}
\arguments{
\item{data}{The raw data from BioSense on which you will do the invalid Diagnosis_Type checks.}
}
\value{
A list of two data frames: examples and summary for invalid Diagnosis_Type.
}
\description{
This function will generate two data frames: first, a frame to be used later to extract invalid examples from;
second, a frame that contains facility-level summaries for counts and percentages of invalid Diagnosis_Type.
}
\details{
The valid values were taken from the `PHVS_DiagnosisType_HL7_2x_V1.xls` file from 
Public Health Information Network Vocabulary Access and Distribution System value sets
(https://phinvads.cdc.gov/vads/ViewView.action?name=Syndromic%20Surveillance). The package will
be updated as the CDC provides new or different codes that are considered valid or invalid.

You can view the concept codes that are considered valid by calling `data("diagnosis_type")`.
}
