% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discharge_disposition_invalid.R
\name{discharge_disposition_invalid}
\alias{discharge_disposition_invalid}
\title{Getting Invalid for Examples and Summaries for Discharge_Disposition}
\usage{
discharge_disposition_invalid(data)
}
\arguments{
\item{data}{The raw data from BioSense on which you will do the invalid discharge disposition checks.}
}
\value{
A list of two data frames: examples and summary for invalid Discharge_Disposition.
}
\description{
This function will generate two data frames: first, a frame to be used later to extract invalid examples from;
second, a frame that contains facility-level summaries for counts and percentages of invalid Discharge_Disposition.
}
\details{
The valid values were taken from the `PHVS_DischargeDisposition_HL7_2x_V3.xls` file from 
Public Health Information Network Vocabulary Access and Distribution System value sets
(https://phinvads.cdc.gov/vads/ViewView.action?name=Syndromic%20Surveillance). The package will
be updated as the CDC provides new or different codes that are considered valid or invalid.

You can view the concept codes that are considered valid by calling `data("discharge_disposition")`. Note that this
data frame does not include zeros in front of single-digit numbers 1 to 9; these are also added as valid values
in the function.
}
