% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples_invalids.R
\name{examples_invalids}
\alias{examples_invalids}
\title{Generate Examples of Invalid Fields}
\usage{
examples_invalids(i, invalid_examples)
}
\arguments{
\item{i}{A facility ID, as a string.}

\item{invalid_examples}{A list of data frames that are the examples output from the `_invalid` functions.}
}
\value{
A data frame containing the patient-visit IDs and fields for this visit that were invalid.
}
\description{
NOTE: THIS FUNCTION IS NOT MEANT TO BE USED ON ITS OWN. IT IS USED INTERNALLY BY `write_reports`.
This function will return the C_BioSense_ID and field name for patient-visits where the entry was invalid.
}
