% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples_nulls.R
\name{examples_nulls}
\alias{examples_nulls}
\title{Generate Examples of Null Fields}
\usage{
examples_nulls(i, data)
}
\arguments{
\item{i}{A facility ID, as a string.}

\item{data}{The raw data from BioSense on which these examples will be generated.}
}
\value{
A data frame containing the patient-visit IDs and fields for this visit that were null.
}
\description{
This function will return the C_BioSense_ID for visits where required fields are missing, as well as what fields that are missing.
As mentioned in `get_req_nulls`, fields were treated as null if they were null in every single record of the visit, *except* for 
ones that had to be present in every record and thus are done at the record-level: Treating_Facility_ID, Trigger_Event, Message_Control_ID, 
Processing_ID, Version_ID, Message_Profile_ID, Recorded_Date_Time, Message_Date_Time, First_Patient_ID, and Medical_Record_Number.
}
