% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facility_type_invalid.R
\name{facility_type_invalid}
\alias{facility_type_invalid}
\title{Getting Invalid Examples and Summary for Facility_Type}
\usage{
facility_type_invalid(data)
}
\arguments{
\item{data}{The raw data from BioSense on which you will do the invalid facility type checks.}
}
\value{
A list of two data frames: examples and summary for invalid Facility_Type.
}
\description{
This function will generate two data frames: first, a frame to be used later to extract invalid examples from;
second, a frame that contains facility-level summaries for counts and percentages of invalid Facility_Type.
}
\details{
The valid values were taken from the `PHVS_FacilityVisitType_SyndromicSurveillance_V3.xls` file from 
Public Health Information Network Vocabulary Access and Distribution System value sets
(https://phinvads.cdc.gov/vads/ViewView.action?name=Syndromic%20Surveillance). The package will
be updated as the CDC provides new or different codes that are considered valid or invalid.

You can view the concept codes that are considered valid by calling `data("facility_type")`.
}
