% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpid_mrn_invalid.R
\name{fpid_mrn_invalid}
\alias{fpid_mrn_invalid}
\title{Testing if First_Patient_ID is equal to one's Medical_Record_Number}
\usage{
fpid_mrn_invalid(data)
}
\arguments{
\item{data}{The raw data from BioSense on which you will do the mismatch checks.}
}
\value{
A list of two data frames: examples and summary for mismatches from First_Patient_ID (FPID) and Medical_Record_Number (MRN).
}
\description{
This function will generate two data frames: first, a frame to be used later to extract invalid examples from;
second, a frame that contains facility-level summaries for counts and percentages of cases where a patient's
First_Patient_ID does not match their Medical_Record_Number.
}
