% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gender_invalid.R
\name{gender_invalid}
\alias{gender_invalid}
\title{Getting Invalid Examples and Summaries for Administrative_Sex}
\usage{
gender_invalid(data)
}
\arguments{
\item{data}{The raw data from BioSense on which you will do the invalid Administrative_Sex checks.}
}
\value{
A list of two data frames: examples and summary for Administrative_Sex.
}
\description{
This function will generate two data frames: first, a frame to be used later to extract invalid examples from;
second, a frame that contains facility-level summaries for counts and percentages of invalid Administrative_Sex.
}
\details{
The valid values were taken from the `"PHVS_Gender_SyndromicSurveillance_V1.xls"` file from 
Public Health Information Network Vocabulary Access and Distribution System value sets
(https://phinvads.cdc.gov/vads/ViewView.action?name=Syndromic%20Surveillance). The package will
be updated as the CDC provides new or different codes that are considered valid or invalid.

You can view the concept codes that are considered valid by calling `data("gender")`.
}
