% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_invalids.R
\name{get_all_invalids}
\alias{get_all_invalids}
\title{Getting Invalid Summaries}
\usage{
get_all_invalids(data)
}
\arguments{
\item{data}{The raw data from BioSense on which you will do the invalid checks.}
}
\value{
A data frame that contains the results from all of the `_invalid` functions.
}
\description{
This function will call upon each of the functions that end with `_invalid` and combine the results from all into one data frame. All
invalid checks were done at the patient-visit-level: If *one* message in a patient-visit was found to be invalid, then the whole patient visit
was returned as invalid.
}
