% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_req_nulls.R
\name{get_req_nulls}
\alias{get_req_nulls}
\title{Getting Nulls Summaries for Required Fields}
\usage{
get_req_nulls(data)
}
\arguments{
\item{data}{The raw data from BioSense on which you will do the null checks.}
}
\value{
A summary data frame that lists counts and percentages for null fields, summarized at facility-level.
}
\description{
This will generate percent and counts of null message, by facility. All are calculated on a once-per-visit basis, except for:
Treating_Facility_ID, Trigger_Event, Message_Control_ID, Processing_ID, Version_ID, Message_Profile_ID, Event_Type_Code, 
Recorded_Date_Time, Message_Date_Time, First_Patient_ID, Medical_Record_Number. These fields are required every single message, so
null counts are based on messages. The rest are required only once per-visit, so the whole visit is considered null *only* 
if *every* record for that field from that visit was null.
}
\details{
The distinctions between optional vs. invalid and per visit vs. every message were made based off of our interpretation of the
PHIN guide as well as the needs of the Kansas Department of Health and Environment Syndromic Surveillance.
}
