% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_facility_summary.R
\name{one_facility_summary}
\alias{one_facility_summary}
\title{Extract Summary for One Facility}
\usage{
one_facility_summary(getsummary, i)
}
\arguments{
\item{getsummary}{A summary from `get_req_nulls`, `get_opt_nulls`, or `get_all_invalids`.}

\item{i}{A facility ID, as a character string.}
}
\value{
The information for only the given facility from the given table, transposed and formatted for `write_reports` to place it in a 
facility summary Excel workbook.
}
\description{
This function takes a summary from `get_req_nulls`, `get_opt_nulls`, and `get_all_invalids` and returns the information for only one facility.
}
