% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_class_invalid.R
\name{patient_class_invalid}
\alias{patient_class_invalid}
\title{Getting Invalid Examples and Summaries for Patient_Class_Code}
\usage{
patient_class_invalid(data)
}
\arguments{
\item{data}{The raw data from BioSense on which you will do the invalid patient class checks.}
}
\value{
A list of two data frames: examples and summary for invalid Patient_Class_Code.
}
\description{
This function will generate two data frames: first, a frame to be used later to extract invalid examples from;
second, a frame that contains facility-level summaries for counts and percentages of invalid Patient_Class_Code.
}
\details{
The valid values were taken from the `PHVS_PatientClass_SyndromicSurveillance_V2.xls` file from 
Public Health Information Network Vocabulary Access and Distribution System value sets
(https://phinvads.cdc.gov/vads/ViewView.action?name=Syndromic%20Surveillance). The package will
be updated as the CDC provides new or different codes that are considered valid or invalid.

You can view the concept codes that are considered valid by calling `data("patient_class")`.
}
