% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{pull_data}
\alias{pull_data}
\title{Pulling Data}
\usage{
pull_data(username, password, table, mft, start, end)
}
\arguments{
\item{username}{Your BioSense username, as a string. This is the same username you may use to log into RStudio or Adminer.}

\item{password}{Your BioSense password, as a string. This is the same password you may use to log into RStudio or Adminer.}

\item{table}{The table that you want to retrieve the data from, as a string.}

\item{mft}{The MFT (master facilities table) from where the facility names will be retrieved, as a string.}

\item{start}{The start date time that you wish to begin pulling data from, as a string.}

\item{end}{The end data time that you wish to stop pulling data from, as a string.}
}
\value{
A list of two data frames: first, raw data of all columns from the `table` that was called; second, just the Facility_Name
and C_Biosense_Facility_ID from the MFT table.
}
\description{
This will pull down data from NSSP's BioSense Platform. It assumes you have a valid connection to BioSense_Platform log-in.
}
