% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoking_status_invalid.R
\name{smoking_status_invalid}
\alias{smoking_status_invalid}
\title{Getting Invalid Examples and Summaries for Smoking_Status_Code}
\usage{
smoking_status_invalid(data)
}
\arguments{
\item{data}{The raw data from BioSense on which you will do the invalid smoking status checks.}
}
\value{
A list of two data frames: examples and summary for invalid Smoking_Status_Code.
}
\description{
This function will generate two data frames: first, a frame to be used later to extract invalid examples from;
second, a frame that contains facility-level summaries for counts and percentages of invalid Smoking_Status_Code.
}
\details{
The valid values were taken from the `PHVS_SmokingStatus_MU_V2.xls` file from 
Public Health Information Network Vocabulary Access and Distribution System value sets
(https://phinvads.cdc.gov/vads/ViewView.action?name=Syndromic%20Surveillance). The package will
be updated as the CDC provides new or different codes that are considered valid or invalid.

You can view the concept codes that are considered valid by calling `data("smoking_status")`.
}
