% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statewide.R
\name{statewide}
\alias{statewide}
\title{Generate State-Level Summaries}
\usage{
statewide(data, state_req_nulls, state_opt_nulls, state_invalids)
}
\arguments{
\item{data}{The raw data from BioSense on which you will generate the statewide summary.}

\item{state_req_nulls}{A summary returned from the `get_req_nulls` function.}

\item{state_opt_nulls}{A summary returned from the `get_opt_nulls` function.}

\item{state_invalids}{A summary returned from the `get_all_invalids` function.}
}
\value{
A list of three two-row data frames, each being counts and percentages for required nulls, optional nulls, and invalid entries.
}
\description{
All of the other functions group by facility and then often group by patient visit. This function generates a true statewide
(or table-wide, if the table you are calling is not the entire state) summary for counts and percents for each null and invalid check.
The same message-level or patient-visit-level rules are used as are described in `get_all_invalids`, `get_opt_nulls` and `get_req_nulls`.
}
