% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_facility.R
\name{write_facility}
\alias{write_facility}
\title{Write NSSP BioSense Platform Data Quality Summary Reports for One Facility}
\usage{
write_facility(username, password, table, mft, start, end, facility,
  directory = "", nexamples = 0)
}
\arguments{
\item{username}{Your BioSense username, as a string. This is the same username you may use to log into RStudio or Adminer.}

\item{password}{Your BioSense password, as a string. This is the same password you may use to log into RStudio or Adminer.}

\item{table}{The table that you want to retrieve the data from, as a string.}

\item{mft}{The MFT (master facilities table) from where the facility name will be retrieved, as a string.}

\item{start}{The start date time that you wish to begin pulling data from, as a string.}

\item{end}{The end data time that you wish to stop pulling data from, as a string.}

\item{facility}{The C_Biosense_Facility_ID for the facility that you wish to generate and write the report for.}

\item{directory}{The directory where you would like to write the reports to (i.e., "~/Documents/MyReports"), as a string.}

\item{nexamples}{An integer number of examples you would like for each type of invalid or null field in the examples workbooks for each facility.
This defaults to 0, which will not generate these example workbooks.}
}
\description{
This function is a lightweight version of the `write_reports` function. It will generate summary and example workbooks, but only for
one specified facility. The first summary workbook shows percents and counts of nulls and invalids, while the examples workbook
generates de tailed information on records and visits that are null or invalid.
}
