% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_reports_local.R
\name{write_reports_local}
\alias{write_reports_local}
\title{Write NSSP BioSense Platform Data Quality Summary Reports, Using Local Data}
\usage{
write_reports_local(data, fnames, directory = "", nexamples = 0)
}
\arguments{
\item{data}{A data frame of raw data from the BioSense Platform.}

\item{fnames}{A data frame that includes two variables: C_Biosense_Facility_ID and Facility_Name. Make sure there is NO repeat C_Biosense_Facility_ID numbers.}

\item{directory}{The directory where you would like to write the reports to (i.e., "~/Documents/MyReports"), as a string.}

\item{nexamples}{An integer number of examples you would like for each type of invalid or null field in the examples workbooks for each facility.
This defaults to 0, which will not generate these example workbooks.}
}
\description{
The `write_reports` function will do a data pull using RODBC. In case you do not have a connection to BioSense_Platform through RODBC, you can
use this function to run the reports on data that are loaded to your current R session. That way, you can load in files yourself if you need to write
reports for data have already been pulled or been pulled down using other means than the RODBC package.
}
