% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/va_lag.R
\name{va_lag}
\alias{va_lag}
\title{Visit-Arrival Lag}
\usage{
va_lag(data, names, offset)
}
\arguments{
\item{data}{The raw data from BioSense on which you will do these summaries.}

\item{names}{A table of Facility_Name and C_Biosense_Facility_ID which taken from the MFT using the `pull_data` function.}

\item{offset}{The number of hours you wish to offset Arrived_Date_Time (which is in UTC). The offset value is how far off your local time zone is from UTC. 
For example, the Central Time Zone would set this to 5 or 6, depending on if it is daylight savings or not. This value should be an integer.}
}
\value{
A data frame with three columns: facility name, facility ID, and average lag between patient visit and first record arrival.
}
\description{
Calculates the average time lag between when: a patient arrives and the first record for their visit is received by the NSSP BioSense Platform.
A summary is returned at the facility level.
}
